+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 
+## Added
+- support docxf and oform formats
+- media types ONLYOFFICE Form and ONLYOFFICE MasterForm
+
 ## Chaged
 - fixed the problem when the file was saved empty if there was an error getting the file from the document server (https://www.drupal.org/project/onlyoffice/issues/3362004)